Adding JUnits for metadata file generation
